수치 다 float으로 바꿈

+++ Neo/Shadow In The Shade/Assets/01.Scripts/Enemy/Boss/Boss_Dice.cs
@@ -78,7 +78,7 @@ public class Boss_Dice : Enemy
         base.CheckHP();
     }
 
-    public override void GetHit(int damage)
+    public override void GetHit(float damage)
     {
         base.GetHit(damage);
     }
