+++ Neo/Shadow In The Shade/Assets/01.Scripts/Enemy/Slime/Slime_Beaker.cs
@@ -123,7 +123,7 @@ public class Slime_Beaker : Enemy, ITacklable
     }
 
 
-    public override void GetHit(int damage)
+    public override void GetHit(float damage)
     {
 
         base.GetHit(damage);
